Extracted the function for adding tokens to a util function and not server action to secure it
